Creating login function on login.php
So before any user is able to log in, the code had to be able to
understand existing records saved within the database - logins, under -
users. For this i had to create a variable $records, in corrosponding to
$conn which should be able to select the required data from the database
table. Subsequently, I also used the PDO FETCH constant, which will
essentially fetch the record off the table that macthes the entered
details in the initial registeration process.