Using sessions within the coding
I had to use session functions to ensure logged in users can stay logged
in throughout browsing the pages. The function, "seesion_start();" is
placed at the very top of the code compiler on all suitable pages. Best
Example is on the 'logout.php' page as i was using most types of session
fucntions.