Secret Page...
Once the user has successfully signed in, they get directed to a so
called 'secret page' which is once only available whenever one has
logged in. To achieve this it was rather simple. only had to change the
location, under '("Location: /");' in the 'login.php' page to
'("Location: loginp.php");' which basically redirects the logged in user
to 'loginp.php'.